New data format, breaks compatibility!
Everything is now properly encrypted or hashed on the archive side. The only usable information remote nodes have now is an approximation of file sizes after compression and a hash of the file names.

Changed the commands to make it easier to use. It's now possible to add a node offline if we know its public key. No need to have the server nodes somehow connect to the client just to add them anymore.

The code is now much simpler and more robust. We now don't blindly trust that remote nodes won't send garbage when deserializing, even if the remote node is trusted.